fix: fix to Research submenus and IPR shifted to Institute
