+++ i18n/en.ts
@@ -559,6 +559,7 @@ const text: Translations = {
     },
     cells: {
       title: 'Cells',
+      headingTitle : 'Institute Cells',
       cell: 'cell',
       iic: {
         title: 'Institution’s Innovation Council',
@@ -584,6 +585,9 @@ const text: Translations = {
           'IKS Cell is an innovative cell established in 2022 in the Institute. It aims to promote interdisciplinary research on all aspects of Indian Knowledge Systems, preserve and disseminate IKS for further research and societal applications. The cell will actively engage in spreading the rich heritage of our country and traditional knowledge in fields such as Psychology, Basic Sciences, Engineering & Technology, Arts and Literature, Agriculture, Architecture, and more.',
         iksTeam: 'IKS Team',
       },
+      ipr:{
+        title: 'Intellectual Property Rights',
+      }
     },
   },
   NotFound: {
