+++ i18n/hi.ts
@@ -509,6 +509,7 @@ const text: Translations = {
     },
     cells: {
       title: 'सेल्स',
+      headingTitle: 'संस्थान सेल्स',
       cell: 'सेल',
       iic: {
         title: 'संस्थान नवाचार परिषद',
@@ -533,6 +534,9 @@ const text: Translations = {
         description:
           'आईकेएस प्रकोष्ठ एक नवाचारी इकाई है जिसे वर्ष 2022 में संस्थान में स्थापित किया गया। इसका उद्देश्य भारतीय ज्ञान प्रणाली के सभी पहलुओं पर अंतःविषयक शोध को प्रोत्साहित करना, IKS को संरक्षित करना और आगे के शोध तथा सामाजिक अनुप्रयोगों के लिए उसका प्रसार करना है। यह प्रकोष्ठ हमारे देश की समृद्ध धरोहर और पारंपरिक ज्ञान को मनोविज्ञान, मूलभूत विज्ञान, अभियंत्रण एवं प्रौद्योगिकी, कला और साहित्य, कृषि, वास्तुकला आदि क्षेत्रों में सक्रिय रूप से फैलाने का कार्य करेगा।',
         iksTeam: 'आईकेएस टीम',
+      },
+      ipr: {
+        title: 'बौद्धिक संपदा अधिकार',
       },
     },
   },
